Split layer-specific tests to separate modules.
